Add ability to create a slot for a hex-headed bolt
Add ability to create a slot for a hex-headed bolt.  This requires a
version of cq-queryabolt that includes support for hex-headed bolts.

It:

- adds helper functions that generate sketches for the slots along
  which the shaft and head slide; and
- provides a method in the Workspace mixin that creates the bolt
  slot.

Natural orientation is looking down on the bottom of a
cube (`faces("<Z")`, with the length component running along the Y
axis), which creates the head-slot along the bottom face and cuts the
shaft slot all the way through the solid.  In other words, it cuts in
the -Z direction relative to the local coordinate system.